test(postage-usage): share integration-test scaffolding
Hoist the bucket depth, batch id, owner, and `Shallow` spec helpers
duplicated across the snapshot, vector, issuer, and seal test binaries into
`tests/common/mod.rs`, following the manifest precedent.
